Applying closed library strategy moved to separate place
