+++ src/main/java/org/folio/circulation/resources/CheckOutByBarcodeResource.java
@@ -4,7 +4,6 @@ import io.vertx.core.http.HttpClient;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
-import org.folio.circulation.domain.CalendarRepository;
 import org.folio.circulation.domain.Item;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.LoanAndRelatedRecords;
@@ -17,8 +16,7 @@ import org.folio.circulation.domain.User;
 import org.folio.circulation.domain.UserRepository;
 import org.folio.circulation.domain.policy.LoanPolicy;
 import org.folio.circulation.domain.policy.LoanPolicyRepository;
-import org.folio.circulation.domain.policy.library.ClosedLibraryStrategy;
-import org.folio.circulation.domain.policy.library.ClosedLibraryStrategyUtils;
+import org.folio.circulation.domain.policy.library.ClosedLibraryStrategyService;
 import org.folio.circulation.domain.representations.CheckOutByBarcodeRequest;
 import org.folio.circulation.domain.representations.LoanProperties;
 import org.folio.circulation.domain.validation.AlreadyCheckedOutValidator;
@@ -33,7 +31,6 @@ import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.CreatedJsonHttpResult;
 import org.folio.circulation.support.HttpResult;
 import org.folio.circulation.support.ItemRepository;
-import org.folio.circulation.support.PeriodUtil;
 import org.folio.circulation.support.RouteRegistration;
 import org.folio.circulation.support.WritableHttpResult;
 import org.folio.circulation.support.http.server.WebContext;
@@ -41,7 +38,6 @@ import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
 import org.joda.time.format.ISODateTimeFormat;
 
-import java.util.Optional;
 import java.util.UUID;
 
 import static java.util.concurrent.CompletableFuture.completedFuture;
@@ -66,17 +62,18 @@ public class CheckOutByBarcodeResource extends Resource {
 
     JsonObject request = routingContext.getBodyAsJson();
 
-    final JsonObject loan = new JsonObject();
-    loan.put("id", UUID.randomUUID().toString());
+    final JsonObject loanJson = new JsonObject();
+    loanJson.put("id", UUID.randomUUID().toString());
 
-    copyOrDefaultLoanDate(request, loan);
+    copyOrDefaultLoanDate(request, loanJson);
 
     final String itemBarcode = request.getString(CheckOutByBarcodeRequest.ITEM_BARCODE);
     final String userBarcode = request.getString(CheckOutByBarcodeRequest.USER_BARCODE);
     final String proxyUserBarcode = request.getString(CheckOutByBarcodeRequest.PROXY_USER_BARCODE);
     final String checkoutServicePointId = request.getString(CheckOutByBarcodeRequest.SERVICE_POINT_ID);
 
-    loan.put(LoanProperties.CHECKOUT_SERVICE_POINT_ID, checkoutServicePointId);
+    loanJson.put(LoanProperties.CHECKOUT_SERVICE_POINT_ID, checkoutServicePointId);
+    Loan loan = Loan.from(loanJson);
 
     final Clients clients = Clients.create(context, client);
 
@@ -85,7 +82,7 @@ public class CheckOutByBarcodeResource extends Resource {
     final RequestQueueRepository requestQueueRepository = RequestQueueRepository.using(clients);
     final LoanRepository loanRepository = new LoanRepository(clients);
     final LoanPolicyRepository loanPolicyRepository = new LoanPolicyRepository(clients);
-    final CalendarRepository calendarRepository = new CalendarRepository(clients);
+    final ClosedLibraryStrategyService strategyService = ClosedLibraryStrategyService.using(clients, loan.getLoanDate());
 
     final ProxyRelationshipValidator proxyRelationshipValidator = new ProxyRelationshipValidator(
       clients, () -> failure(
@@ -122,7 +119,7 @@ public class CheckOutByBarcodeResource extends Resource {
 
     final LoanRepresentation loanRepresentation = new LoanRepresentation();
 
-    completedFuture(HttpResult.succeeded(new LoanAndRelatedRecords(Loan.from(loan))))
+    completedFuture(HttpResult.succeeded(new LoanAndRelatedRecords(loan)))
       .thenApply(servicePointOfCheckoutPresentValidator::refuseCheckOutWhenServicePointIsNotPresent)
       .thenCombineAsync(userRepository.getUserByBarcode(userBarcode), this::addUser)
       .thenCombineAsync(userRepository.getProxyUserByBarcode(proxyUserBarcode), this::addProxyUser)
@@ -138,10 +135,7 @@ public class CheckOutByBarcodeResource extends Resource {
       .thenComposeAsync(r -> r.after(loanPolicyRepository::lookupLoanPolicy))
       .thenApply(itemIsNotLoanableValidator::refuseWhenItemIsNotLoanable)
       .thenApply(r -> r.next(this::calculateDefaultInitialDueDate))
-      .thenComposeAsync(r -> r.after(calendarRepository::lookupPeriod))
-      .thenApply(r -> r.next(this::applyCLDDM))
-      .thenComposeAsync(r -> r.after(calendarRepository::lookupPeriodForFixedDueDateSchedule))
-      .thenApply(r -> r.next(this::applyFixedDueDateLimit))
+      .thenComposeAsync(r -> r.after(strategyService::applyCLDDM))
       .thenComposeAsync(r -> r.after(requestQueueUpdate::onCheckOut))
       .thenComposeAsync(r -> r.after(updateItem::onCheckOut))
       .thenComposeAsync(r -> r.after(loanRepository::createLoan))
@@ -159,56 +153,6 @@ public class CheckOutByBarcodeResource extends Resource {
         loanAndRelatedRecords.getLoan().changeDueDate(dueDate);
         return loanAndRelatedRecords;
       });
-  }
-
-  private HttpResult<LoanAndRelatedRecords> applyCLDDM(LoanAndRelatedRecords relatedRecords) {
-    if (relatedRecords.getInitialDueDateDays() == null) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-    DateTimeZone timeZone = relatedRecords.getTimeZone();
-    ClosedLibraryStrategy strategy =
-      ClosedLibraryStrategyUtils.determineClosedLibraryStrategy(
-        relatedRecords.getLoanPolicy(),
-        relatedRecords.getLoan().getLoanDate(),
-        timeZone);
-
-    DateTime dueDate = relatedRecords.getLoan().getDueDate();
-    HttpResult<DateTime> calculateDueDate =
-      strategy.calculateDueDate(dueDate, relatedRecords.getInitialDueDateDays());
-    return calculateDueDate.next(date -> {
-      relatedRecords.getLoan().changeDueDate(date);
-      return HttpResult.succeeded(relatedRecords);
-    });
-  }
-
-  private HttpResult<LoanAndRelatedRecords> applyFixedDueDateLimit(LoanAndRelatedRecords relatedRecords) {
-    if (relatedRecords.getFixedDueDateDays() == null) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-    final Loan loan = relatedRecords.getLoan();
-    final LoanPolicy loanPolicy = relatedRecords.getLoanPolicy();
-    final DateTime loanDate = relatedRecords.getLoan().getLoanDate();
-    final DateTime dueDate = relatedRecords.getLoan().getDueDate();
-
-    Optional<DateTime> optionalDueDateLimit = loanPolicy.getFixedDueDateSchedules()
-      .findDueDateFor(loan.getLoanDate());
-    if (!optionalDueDateLimit.isPresent()) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-    DateTime dueDateLimit = optionalDueDateLimit.get();
-    if (!PeriodUtil.isAfterDate(loan.getDueDate(), dueDateLimit)) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-
-    ClosedLibraryStrategy strategy =
-      ClosedLibraryStrategyUtils.determineStrategyForMovingBackward(
-        loanPolicy, loanDate, relatedRecords.getTimeZone());
-    HttpResult<DateTime> calculatedDate =
-      strategy.calculateDueDate(dueDate, relatedRecords.getFixedDueDateDays());
-    return calculatedDate.next(date -> {
-      relatedRecords.getLoan().changeDueDate(date);
-      return HttpResult.succeeded(relatedRecords);
-    });
   }
 
   private void copyOrDefaultLoanDate(JsonObject request, JsonObject loan) {
