+++ src/main/java/org/folio/circulation/resources/RenewalResource.java
@@ -4,28 +4,22 @@ import io.vertx.core.http.HttpClient;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.Router;
 import io.vertx.ext.web.RoutingContext;
-import org.folio.circulation.domain.CalendarRepository;
 import org.folio.circulation.domain.Loan;
 import org.folio.circulation.domain.LoanAndRelatedRecords;
 import org.folio.circulation.domain.LoanRenewalService;
 import org.folio.circulation.domain.LoanRepository;
 import org.folio.circulation.domain.LoanRepresentation;
 import org.folio.circulation.domain.UserRepository;
-import org.folio.circulation.domain.policy.LoanPolicy;
 import org.folio.circulation.domain.policy.LoanPolicyRepository;
-import org.folio.circulation.domain.policy.library.ClosedLibraryStrategy;
-import org.folio.circulation.domain.policy.library.ClosedLibraryStrategyUtils;
+import org.folio.circulation.domain.policy.library.ClosedLibraryStrategyService;
 import org.folio.circulation.domain.representations.LoanResponse;
 import org.folio.circulation.support.Clients;
 import org.folio.circulation.support.HttpResult;
 import org.folio.circulation.support.ItemRepository;
-import org.folio.circulation.support.PeriodUtil;
 import org.folio.circulation.support.RouteRegistration;
 import org.folio.circulation.support.http.server.WebContext;
 import org.joda.time.DateTime;
-import org.joda.time.DateTimeZone;
 
-import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
 public abstract class RenewalResource extends Resource {
@@ -52,10 +46,10 @@ public abstract class RenewalResource extends Resource {
     final ItemRepository itemRepository = new ItemRepository(clients, true, true);
     final UserRepository userRepository = new UserRepository(clients);
     final LoanPolicyRepository loanPolicyRepository = new LoanPolicyRepository(clients);
-    final CalendarRepository calendarRepository = new CalendarRepository(clients);
 
     final LoanRepresentation loanRepresentation = new LoanRepresentation();
     final LoanRenewalService loanRenewalService = LoanRenewalService.using(clients);
+    final ClosedLibraryStrategyService strategyService = ClosedLibraryStrategyService.using(clients, DateTime.now());
 
     //TODO: Validation check for same user should be in the domain service
 
@@ -63,60 +57,11 @@ public abstract class RenewalResource extends Resource {
       .thenApply(r -> r.map(LoanAndRelatedRecords::new))
       .thenComposeAsync(r -> r.after(loanPolicyRepository::lookupLoanPolicy))
       .thenApply(r -> r.next(loanRenewalService::renew))
-      .thenComposeAsync(r -> r.after(calendarRepository::lookupPeriod))
-      .thenApply(r -> r.next(this::applyCLDDM))
-      .thenComposeAsync(r -> r.after(calendarRepository::lookupPeriodForFixedDueDateSchedule))
-      .thenApply(r -> r.next(this::applyFixedDueDateLimit))
+      .thenComposeAsync(r -> r.after(strategyService::applyCLDDM))
       .thenComposeAsync(r -> r.after(loanRepository::updateLoan))
       .thenApply(r -> r.map(loanRepresentation::extendedLoan))
       .thenApply(LoanResponse::from)
       .thenAccept(result -> result.writeTo(routingContext.response()));
-  }
-
-  private HttpResult<LoanAndRelatedRecords> applyCLDDM(LoanAndRelatedRecords relatedRecords) {
-    if (relatedRecords.getInitialDueDateDays() == null) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-    ClosedLibraryStrategy strategy =
-      ClosedLibraryStrategyUtils.determineClosedLibraryStrategy(
-        relatedRecords.getLoanPolicy(), DateTime.now(), relatedRecords.getTimeZone());
-
-    DateTime dueDate = relatedRecords.getLoan().getDueDate();
-    HttpResult<DateTime> calculateDueDate =
-      strategy.calculateDueDate(dueDate, relatedRecords.getInitialDueDateDays());
-    return calculateDueDate.next(date -> {
-      relatedRecords.getLoan().changeDueDate(date);
-      return HttpResult.succeeded(relatedRecords);
-    });
-  }
-
-  private HttpResult<LoanAndRelatedRecords> applyFixedDueDateLimit(LoanAndRelatedRecords relatedRecords) {
-    if (relatedRecords.getFixedDueDateDays() == null) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-    final Loan loan = relatedRecords.getLoan();
-    final LoanPolicy loanPolicy = relatedRecords.getLoanPolicy();
-    final DateTime dueDate = relatedRecords.getLoan().getDueDate();
-
-    Optional<DateTime> optionalDueDateLimit = loanPolicy.getFixedDueDateSchedules()
-      .findDueDateFor(loan.getLoanDate());
-    if (!optionalDueDateLimit.isPresent()) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-    DateTime dueDateLimit = optionalDueDateLimit.get();
-    if (!PeriodUtil.isAfterDate(loan.getDueDate(), dueDateLimit)) {
-      return HttpResult.succeeded(relatedRecords);
-    }
-    DateTimeZone timeZone = relatedRecords.getTimeZone();
-    ClosedLibraryStrategy strategy =
-      ClosedLibraryStrategyUtils.determineStrategyForMovingBackward(
-        loanPolicy, DateTime.now(), timeZone);
-    HttpResult<DateTime> calculatedDate =
-      strategy.calculateDueDate(dueDate, relatedRecords.getFixedDueDateDays());
-    return calculatedDate.next(date -> {
-      relatedRecords.getLoan().changeDueDate(date);
-      return HttpResult.succeeded(relatedRecords);
-    });
   }
 
   protected abstract CompletableFuture<HttpResult<Loan>> findLoan(
